Initial Commit:
An application that allows a user to generate a song (or individual
parts of a song). The generated song itself may not be great, but it can
serve as at least serve as a starting point or inspiration for a song.

Initial commit includes the following functionality:
-A basic screen with a single button and a single text display.
-Pressing the button will generate and display a simple song structure
	-The song structure consists of Verse, Chorus, and Bridges
	-The order of these is determined based on some very basic probability
stuff (that needs some improvement)