Committing a bunch of changes:
Probably the biggest is the addition of MIDI support. Generating a song
now creates a temporary MIDI file, which can then be played via the
built-in MediaPlayer. As it is currently, it plays the verse chords.

Speaking of the verse chords, I added a new method the generate a
(hopefully) better chord progression. It should eliminate repeated
chords. It could still use osme improvements... I'm just going for
incremental changes right now.

Back to MIDI for a minute... this uses an external library (may need to
do some work to make sure licensing is handled appropriately). Including
the .jar. The MidiManager handles most of the Midi processing.

Created some helper classes (Utils and SongStructure) to handle some of
the common operations and constants.

One goal I have is to abstract things enough so that the SongWriter and
MidiManager classes (along with the helper classes) could be used for
Java in general, so I could theoretically create a desktop app that does
a similar thing. I need to learn some Java GUI stuff first, but Android
makes this pretty easy for now.

Anyway, that was pretty long for a commit message...

Oh, also removing some (hopefully) unnecessary files