I'm pretty excited about this one...
Added the ability to save the song to an external MIDI file (it will be
stored in the main music directory on external storage). It can be
played back using various media players, and can also be viewed and
edited in other software (such as MuseScore, which I am using on my PC).
This is useful for viewing it in musical notation.

As part of this, I implemented a dialog fragment that allows the user to
specify a file name to save the MIDI under (the .mid extension is
automatically added). Ideally, this fragment could be reused.

One other thing I learned... android dialogs are a good bit different
than MFC dialogs... Anyway, I digress.