A fair bit of refactoring. Added classes for Chord progressions, chord patterns (sort of like sub-progressions), and Notes (Notes are still not being used yet). This should help with generation and management of these items, particularly when it  comes to using the SongWriter.
As part of this, I added a pretty serious hack to restore state when the
activity is recreated from a saved instance. What I should probably do
is have the objects be serializable (or parseable), or at least be
little more careful about reading and writing them on save/restore.

This check in is mostly to give me a relatively stable point to come
back to as I plan to continue refactoring the code to be a little more
OO and organized.