Make some changes so that melodies no longer just consist of quarter notes. This makes songs at least marginally more unique and interesting.
Also made some more changes to introduce more ways to have a chance of
repeating a part. This will likely continue, since repetition seems to
be a big part in music. It helps unify a song (as opposed to having it
sound so random, as most songs that this is generating sound now).