New features: -arguably large improvements to song generation. Individual sections of a progression are repeated more, but they also have a chance to introduce variations in the melody so that it is more interesting. -added methods to generate melody variations -added methods to handle cadences to make various musical sections have more "resolution" -added chorus to playback and chorus notes to display
Bugs:
-fixed instrument picking (was off by one). This meant some of the
instruments needed to be changed in baseInstruments (because they
weren't what I thought they were)
-removed spinners from layout (since they weren't doing anything anyway)
-fixed a bug where I was assuming that assigning an ArrayList to a
variable would assign by value as opposed to by reference (which is not
the case, by the way). This was causing me to have several variables
refer to the same ArrayLst (when I was assuming they were individual
copies), so when I modified one, it modified them all.