A few enhancements:
-Changed display to use spinners and edit fields to show properties of
the song (instruments, tempo, key, etc). Eventually I will use these to
allow a user to specify these values to generate specific types of songs
(or... hopefully... be able to modify a generated song on the fly!...
Ok, some things like time signature will difficult, if not impossible to
modify on the fly)
-A couple of utility functions were added to facilitate the above

-Added ability to serialize the song so it can be properly restored when
the activity is recreated (WOOT!). This will hopefully also set up the
ability to save songs and reload them in the app later (as opposed to
just exporting to MIDI)
-This uses the GSON library to convert the song to a JSON string (this
is an open source library. Need to add section in the app settings or
something to display open source licenses)

-A few other changes to start applying a new style