Some initial messing around with alternative displays for the song parts.
The current implementation show the verse. It is a HorizontalScrollView
inside the main ScrollView. This houses a linearlayout, into which I
place buttons (which I am calling "songBlocks" for now). Each songBlock
shows 1 measure, with the chord scale degree and melody notes for that
measure (since each measure only has one chord right now)

The idea is that this will be a little easier to understand what these
things mean, and to eventually allow user interaction (which is why they
are buttons, for now) so that users can control/modify these. 

This is only intended to be a rough version of the display (with no
interaction) for now. I am still trying to decide if this is really the
direction I want to go (I am more thinking I want some kind of grid
structure to put these in, but none of the existing android
widgets/views does quite what I want).