Allow the synth to continue playing on app restart:
Use the newly added FluidDroidSynthFragment to persist the synth and its
state on app restart (such as when rotating the screen). Also, since the
synth now handles the audio focus, remove that code from here.

Also, attempt to make sure the support library is in synch with the one
in the FluidDroid library

NOTE TO FUTURE BRYAN: I removed a lot of commented out code that has
been just sitting around for a long time. If you decide you actually
wanted any of that, you can find it here.