Initial changes for displaying chord names in song blocks.
Currently, it now displays the name of the pitch at the root of the
chord. If it happens to be a sharp or flat pitch, it shows the
corresponding flat or sharp note, respectively (i.e. it shows C# and
Db). Eventually it would be nice just to show the correct one for the
scale, but there is much more work to be done for that.

Additionally, this does not yet show the tonality (I think that is the
right word...?) of the chord (whether it is major or minor). Still
working on that as well...