Initial changes for about page (which includes open source licenses)
The code has been added to switch the about page. The page is accessed
through the overflow menu. Right now, the about page consists of just a
single text view with a big block of text (and no hyperlinks). Work on
this to make it a bit more attractive, and add links to the open source
license texts.