Some changes to correctly randomly generate song parameters
The way things were before worked fine on newer versions of Android
(presumably because they made onItemSelected not be stupid and get
called all the freaking time...). But on older versions, there was some
funny business going on with syncing the spinner controls with the
song/songWriter. Namely, values were getting improperly overwritten in
this code. Now, we do things a little more carefully, and it seems to
work (at least in the emulator for 3.0).

Also have to update song generator jar to include a few new songWriter
functions.