Update to only support Android 2.3 and later
There is an issue I ran into with the synth writing the soundfont on
Android 2.2 (in an emulator). It seems to be a bug in that version of
Android without a clean workaround (can't write certain files larger
than 1MB). It seems to work fine in 2.3 and later though.

Also, according to the latest distribution numbers, 2.2 only accounts
for 0.8% of the Android market, so it should not be a big deal to not
support it.

Also, remove the specified sizing sizing of the tempo spinner. It
doesn't work very well at all in older version of Android.