Remove parent activity from about activity
There was an issue where going to the about page after generating a song
and using the back button in the title bar would take you back to the
main activity with no song associated with it (which could cause
problems, since a song could be playing when going back and forth
between the activities). Removing the parent/child relationship forces
the user to use the system back button, which will take them back to the
same activity with the correct song info.